updating new test

+++ src/resolvers/Mutations/SingOut.ts
@@ -1,8 +1,14 @@
 import { createRouter } from '../../router/createRouter'
 import * as JWT from '../../auth/jwt'
-
+import * as trpc from '@trpc/server'
 const resolve = async ({ ctx }) => {
-  const jwt = await JWT.getUserToken(ctx)
+  const jwt = JWT.getUserToken(ctx)
+  if (jwt == null) {
+    throw new trpc.TRPCError({
+      code: 'UNAUTHORIZED',
+      message: 'not permitted!',
+    })
+  }
   await ctx.prisma.session.delete({
     where: {
       id: jwt['sessionId'],
